+++ src/rules/element-name.ts
@@ -1,3 +1,4 @@
+import { sliceLocation } from "../context";
 import { TagOpenEvent } from "../event";
 import { Rule, RuleDocumentation, ruleDocumentationUrl } from "../rule";
 
@@ -28,10 +29,11 @@ class ElementName extends Rule {
 		this.on("tag:open", (event: TagOpenEvent) => {
 			const target = event.target;
 			const tagName = target.tagName;
+			const location = sliceLocation(event.location, 1);
 
 			/* check if element is blacklisted */
 			if (this.options.blacklist.indexOf(tagName) >= 0) {
-				this.report(target, `<${tagName}> element is blacklisted`);
+				this.report(target, `<${tagName}> element is blacklisted`, location);
 			}
 
 			/* assume that an element with meta has valid name as it is a builtin
@@ -52,7 +54,11 @@ class ElementName extends Rule {
 			}
 
 			if (!tagName.match(this.pattern)) {
-				this.report(target, `<${tagName}> is not a valid element name`);
+				this.report(
+					target,
+					`<${tagName}> is not a valid element name`,
+					location
+				);
 			}
 		});
 	}